Megre from Master

+++ src/main/java/com/alekseytyan/dao/api/LorryDao.java
@@ -5,5 +5,5 @@ import com.alekseytyan.entity.Lorry;
 import java.util.List;
 
 public interface LorryDao extends AbstractDao<Lorry, String> {
-    List<Lorry> findSuitableLorries(Long weight);
+    List<Lorry> findSuitableLorries(int weight);
 }
